Move webpack tests to subdirectory within tests directory
