+++ test/webpack/test-project/src/concat-files/a/3.js
