+++ test/webpack/test-project/src/concat-files/b/2.js
