+++ test/webpack/test-project/src/index.js
